Listo Signup en Publicar Venta, hace falta corregir algo de back para hacer pruebas.

+++ src/components/FormSignUpVendedora.vue
@@ -14,16 +14,25 @@
           .upfile__small
             h3.upfile__title Foto de perfil
             .upfile__item
+              a.delete(
+                v-show='toggleImageDelete',
+                @click='removeImage') Eliminar
+              span.help(
+                v-if="errorLog.picture"
+              ) {{ errorLog.picture }}
               .upfile__label
                 .upfile__text.i-upload(
                   v-if="mqDesk") Arrastra una foto o
                 .upfile__btn Sube una imagen
               croppa(
-                :width="300",
-                :height="300",
-                :quality="2",
-                placeholder="",
-                :prevent-white-space="true")
+                v-model='picture',
+                :width='300',
+                :height='300',
+                :quality='2',
+                placeholder='',
+                :prevent-white-space='true',
+                @new-image-drawn='handlePicture',
+                @draw='handlePicture')
         .form-section__inner
           p.form__mesagge Ups! Vemos que aún no tienes una cuenta. Te crearemos una cuenta rápidamente para que puedas continuar con la publicación de tu venta. Por favor ingresa los siguientes datos:
           .form__row(
@@ -38,8 +47,7 @@
               type='text',
               data-vv-name='nombre'
             )
-          .form__row(
-            :class='{ "is-danger": errorLog.apellidos }')
+          .form__row(:class='{ "is-danger": errorLog.apellidos }')
             label.form__label(
               for='apellidos') Apellidos
             span.help(
@@ -77,7 +85,7 @@
           fieldset.form__set
             legend.form__legend Nueva dirección
             .form__grid
-              .form__row
+              .form__row(:class='{ "is-danger": errorLog.new_address }')
                 label.form__label(
                   for='new-address') Dirección
                 span.help(
@@ -86,7 +94,7 @@
                   id='new-address'
                   v-model="newAddressData['new_address']"
                   type='text')
-              .form__row
+              .form__row(:class='{ "is-danger": errorLog.new_region }')
                 label.form__label(
                   for='new-address-region') Región
                 span.help(
@@ -98,16 +106,18 @@
                     v-for="region in regions") {{ region }}
 
             .form__grid
-              .form__row
+              .form__row(:class='{ "is-danger": errorLog.new_city }')
                 label.form__label(
                   for='new-address-city') Ciudad
+                span.help(
+                  v-show="errorLog.new_city") {{ errorLog.new_city }}
                 select.form__select(
                   v-model="newAddressData['new_city']")
                   option
                   option(
-                    v-for="citi in cities") {{ citi }}
+                    v-for="city in cities") {{ city }}
 
-              .form__row
+              .form__row(:class='{ "is-danger": errorLog.new_zone }')
                 label.form__label(
                   for='new-address-zone') Comuna
                 span.help(
@@ -146,9 +156,9 @@
 </template>
 
 <script>
-import axios from 'axios'
 import Vue from 'vue'
 import Croppa from 'vue-croppa'
+import userAPI from '@/api/user'
 import userAddressesAPI from '@/api/userAddresses'
 Vue.component('croppa', Croppa.component)
 
@@ -176,27 +186,58 @@ export default {
       },
       newAddressData: {...addressFields},
       infoTexts: {},
-      regionsList: {}
+      regionsList: {},
+      picture: null,
+      pictureURL: null,
+      toggleImageDelete: false
     }
   },
   methods: {
     signUp: function () {
       // console.log(this.$store.get('userAuth'))
-      axios.post('https://prilov.aguayo.co/api/users', {
+      const payload = {
         first_name: this.nombre,
         last_name: this.apellidos,
         email: this.email,
-        password: this.password
-      })
+        password: this.password,
+        picture: this.pictureURL
+      }
+      const modal = {
+        name: 'ModalMessage',
+        parameters: {
+          type: 'preload',
+          title: '¡Gracias! Estamos creando tu cuenta'
+        }
+      }
+      this.$store.dispatch('ui/showModal', modal)
+      userAPI.create(payload)
         .then(response => {
-          console.log(response.data)
-          this.setSuccess()
           localStorage.setItem('token', response.data.api_token)
           localStorage.setItem('userId', response.data.id)
-          this.$store.dispatch('user/setUser', response.data)
+
+          this.$store.dispatch('user/createAddress', this.newAddressData).then(() => {
+            this.$store.dispatch('ui/closeModal')
+            this.$router.push('publicar-venta')
+          }).catch((e) => {
+            console.log(e)
+          })
         })
         .catch(e => {
-          if (e.response.data.errors.exists) this.infoTexts.emailExist = 'Parece que este email ya está siendo usado. ¿Olvidaste tu contraseña?'
+          this.$store.dispatch('ui/closeModal')
+          if (e.response.data.errors.exists) {
+            this.infoTexts.emailExist = 'Parece que este email ya está siendo usado. ¿Olvidaste tu contraseña?'
+            this.validatePassword()
+          } else {
+            const modal = {
+              name: 'ModalMessage',
+              parameters: {
+                type: 'alert',
+                title: '¡Ups! Parece que ocurrió un error',
+                body: Object.values(e.response.data.errors)[0]
+              }
+            }
+            this.$store.dispatch('ui/showModal', modal)
+          }
           this.validatePassword()
         })
     },
@@ -208,6 +249,10 @@ export default {
 
       if (!this.nombre) this.errorLog.nombre = 'Debes ingresar tu nombre'
       if (!this.apellidos) this.errorLog.apellidos = 'Debes ingresar tus apellidos'
+      if (!this.newAddressData.new_address) this.errorLog.new_address = 'Debes ingresar una dirección'
+      if (!this.newAddressData.new_region) this.errorLog.new_region = 'Debes ingresar una región'
+      if (!this.newAddressData.new_city) this.errorLog.new_city = 'Debes ingresar una ciudad'
+      if (!this.newAddressData.new_zone) this.errorLog.new_zone = 'Debes ingresar una comuna'
       if (!this.email) {
         this.errorLog.email = 'Debes ingresar tu email'
       } else {
@@ -231,14 +276,17 @@ export default {
       if (this.password.length < 8) this.errorLog.passwordDetail.push('Tu contraseña debe tener al menos 8 caracteres')
       if (!/[a-zA-Z]/.test(this.password)) this.errorLog.passwordDetail.push('Tu contraseña debe contener al menos una letra')
       if (!/\d+/.test(this.password)) this.errorLog.passwordDetail.push('Tu contraseña debe contener al menos un número')
+      if (!this.picture.hasImage()) this.errorLog.picture = 'Debes cargar una imagen para tu perfil'
     },
-    setSuccess: function () {
-      this.flagSignUp = 'Success'
-      this.$emit('setSuccess')
+    handlePicture: function () {
+      if (this.picture.hasImage()) {
+        this.pictureURL = this.picture.generateDataUrl()
+        this.toggleImageDelete = true
+      }
     },
-    setError: function () {
-      this.flagSignUp = 'Error'
-      this.$emit('setError')
+    removeImage: function (index) {
+      this.toggleImageDelete = false
+      this.picture.remove()
     }
   },
   computed: {
@@ -262,7 +310,6 @@ export default {
     }
   },
   created: function () {
-    this.$store.dispatch('user/loadAddresses')
     const vm = this
     userAddressesAPI.getRegions().then((response) => {
       vm.regionsList = response.data
